src/feature_extractor.py
"""
CLIP-based patch feature extractor for anomaly detection.
Author: Abhay A | github.com/abhay
"""

import torch
import torch.nn.functional as F
import numpy as np
from PIL import Image
from torchvision import transforms
import open_clip


class CLIPPatchExtractor:
    """
    Extracts dense patch-level embeddings using CLIP ViT backbone.
    Supports OpenVINO export for edge deployment (Jetson / CPU).
    """

    def __init__(
        self,
        model_name: str = "ViT-B-16",
        pretrained: str = "openai",
        patch_size: int = 16,
        stride: int = 8,
        device: str = "cuda" if torch.cuda.is_available() else "cpu",
    ):
        self.device = device
        self.patch_size = patch_size
        self.stride = stride

        self.model, _, self.preprocess = open_clip.create_model_and_transforms(
            model_name, pretrained=pretrained
        )
        self.model = self.model.to(device).eval()

        # Standard ImageNet normalization
        self.transform = transforms.Compose([
            transforms.Resize((224, 224)),
            transforms.ToTensor(),
            transforms.Normalize(
                mean=[0.485, 0.456, 0.406],
                std=[0.229, 0.224, 0.225]
            ),
        ])

    @torch.no_grad()
    def extract_global(self, image: Image.Image) -> torch.Tensor:
        """Global image embedding — used for VLM prior alignment."""
        x = self.transform(image).unsqueeze(0).to(self.device)
        feat = self.model.encode_image(x)
        return F.normalize(feat, dim=-1)

    @torch.no_grad()
    def extract_patches(self, image: Image.Image) -> torch.Tensor:
        """
        Dense patch embeddings via sliding window.
        Returns: (N_patches, embed_dim)
        """
        x = self.transform(image).unsqueeze(0).to(self.device)
        # Hook intermediate ViT features for patch-level representation
        patches = self._extract_intermediate_features(x)
        return F.normalize(patches, dim=-1)

    def _extract_intermediate_features(self, x: torch.Tensor) -> torch.Tensor:
        """Extract features from intermediate ViT layers (layers 6+9 avg)."""
        features = []

        def hook_fn(module, input, output):
            # output shape: (B, N_tokens, C) — skip CLS token
            features.append(output[:, 1:, :].detach())

        hooks = []
        target_layers = [6, 9]
        for i, block in enumerate(self.model.visual.transformer.resblocks):
            if i in target_layers:
                hooks.append(block.register_forward_hook(hook_fn))

        with torch.no_grad():
            self.model.encode_image(x)

        for h in hooks:
            h.remove()

        # Average across target layers → (B, N_patches, C)
        stacked = torch.stack(features, dim=0).mean(dim=0)
        return stacked.squeeze(0)  # (N_patches, C)

    def export_to_openvino(self, output_path: str = "clip_encoder.xml"):
        """
        Export encoder to OpenVINO IR for CPU/Jetson inference.
        Achieves ~3-5x speedup on CPU vs PyTorch.
        """
        try:
            from openvino.tools import mo
            import openvino as ov

            dummy = torch.randn(1, 3, 224, 224).to("cpu")
            self.model = self.model.to("cpu")

            torch.onnx.export(
                self.model.visual,
                dummy,
                "/tmp/clip_visual.onnx",
                input_names=["image"],
                output_names=["features"],
                dynamic_axes={"image": {0: "batch"}},
                opset_version=14,
            )

            core = ov.Core()
            model_ov = core.read_model("/tmp/clip_visual.onnx")
            compiled = core.compile_model(model_ov, "CPU")
            ov.save_model(model_ov, output_path)
            print(f"[✓] OpenVINO model saved to {output_path}")
            return compiled

        except ImportError:
            print("[!] OpenVINO not installed. Run: pip install openvino")
