requirements.txt
torch>=2.0.0
torchvision>=0.15.0
open-clip-torch>=2.24.0
numpy>=1.24.0
scikit-learn>=1.3.0
Pillow>=10.0.0
opencv-python>=4.8.0
matplotlib>=3.7.0
tqdm>=4.65.0
gradio>=4.0.0
openai>=1.0.0
requests>=2.31.0

# Optional — edge deployment
# openvino>=2024.0
# tensorrt  # Install via NVIDIA
